pyproject.toml
[tool.poetry]
name = "jwtifypy"
version = "0.1.0"
description = ""
authors = ["LordCode <9999269010dddd@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyjwt = "^2.10.1"


[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.1.1"
pytest = "^8.4.1"
cryptography = "^45.0.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
